+++ doc/changelog.rst
@@ -9,6 +9,11 @@ Unrealeased
   When releasing next version, the "Unreleased" header will be replaced
   with appropriate version header and this help text will be removed.
 
+Admin
+~~~~~
+
+- Add `HexColorWidget` to pick hexadecimal colors on input fields
+
 General
 ~~~~~~~